requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pandas==2.2.3
tenacity==8.2.2
